+++ Elements.Serialization.DXF/src/DxfConverter.cs
@@ -1,6 +1,9 @@
 
 using Elements;
 using Elements.Geometry;
+using Elements.Serialization.DXF.Extensions;
+using IxMilia.Dxf;
+using IxMilia.Dxf.Blocks;
 using IxMilia.Dxf.Entities;
 
 namespace Elements.Serialization.DXF
@@ -24,6 +27,9 @@ namespace Elements.Serialization.DXF
         public double Depth;
     }
 
+    /// <summary>
+    /// Provides information about the model and the drawing configuration. 
+    /// </summary>
     public class DxfRenderContext
     {
         /// <summary>
@@ -35,16 +41,75 @@ namespace Elements.Serialization.DXF
         /// The drawing range that is currently being modeled.
         /// </summary>
         public DrawingRange DrawingRange;
+
     }
 
     /// <summary>
-    /// Interface used during ModelToDxf rendering.
+    /// A class that provides a mechanism for converting elements into DXF
+    /// should inherit this interface.
     /// </summary>
     public interface IRenderDxf
     {
         /// <summary>
-        /// Create a netDxf Entity Object for a given Hypar Element.
+        /// Add a DXF entity to a document
         /// </summary>
-        bool TryToCreateDxfEntity(Element element, DxfRenderContext context, out DxfEntity entity);
+        void TryAddDxfEntity(DxfFile document, Element element, DxfRenderContext context);
+    }
+
+    /// <summary>
+    /// A generic implementation of the IRenderDxf interface.
+    /// </summary>
+    public abstract class DxfConverter<T> : IRenderDxf where T : Element
+    {
+        /// <summary>
+        /// Add a DXF entity to the document for an element.
+        /// </summary>
+        public abstract void TryAddDxfEntity(DxfFile document, T element, DxfRenderContext context);
+
+        /// <summary>
+        /// Add a DXF entity to the document for an element.
+        /// </summary>
+        public void TryAddDxfEntity(DxfFile document, Element element, DxfRenderContext context)
+        {
+            this.TryAddDxfEntity(document, element as T, context);
+        }
+    }
+
+    /// <summary>
+    /// A generic implementation of the IRenderDxf interface for any GeometricElement.
+    /// </summary>
+    public abstract class GeometricDxfConverter<T> : DxfConverter<T> where T : GeometricElement
+    {
+        /// <summary>
+        /// Add a DXF entity to the document for an element.
+        /// </summary>
+        public override void TryAddDxfEntity(DxfFile document, T element, DxfRenderContext context)
+        {
+            // TODO: handle context / drawing range / etc.
+            if (element.Representation == null)
+            {
+                return;
+            }
+            var entities = element.GetEntitiesFromRepresentation();
+            if (element.IsElementDefinition)
+            {
+                var block = new DxfBlock
+                {
+                    Name = element.GetBlockName(),
+                    BasePoint = new DxfPoint(0, 0, 0)
+                };
+                foreach (var e in entities)
+                {
+                    block.Entities.Add(e);
+                }
+                document.Blocks.Add(block);
+                document.BlockRecords.Add(new DxfBlockRecord(block.Name));
+                return;
+            }
+            foreach (var e in entities)
+            {
+                document.Entities.Add(e);
+            }
+        }
     }
 }